sistema_pos/requirements.txt
customtkinter
requests
pandas
pillow
python-dotenv
